refactor: consolidate settings dialog 7-fragment include chain into one file
Merged ui_windows_settings_dialog_{ids,template,state,scroll,paint,handlers,show}.hpp
into a single self-contained ui_windows_settings_dialog.hpp. The fragments were ordered
slices of one implementation included inside a namespace — not independent modules.
The merged file has its own #pragma once, includes, and namespace ui declaration.
