+++ src/ui_windows.hpp
@@ -14,24 +14,7 @@
 #include "app.hpp"
 #include "dialog_style.hpp"
 #include "painting_analog.hpp"
-
-namespace ui {
-using WindowHandle = HWND;
-using FontHandle = HFONT;
-using ThemeHandle = const Theme*;
-using InstanceHandle = HINSTANCE;
-
-namespace settings_dlg {
-#include "ui_windows_settings_dialog_ids.hpp"
-#include "ui_windows_settings_dialog_template.hpp"
-#include "ui_windows_settings_dialog_state.hpp"
-#include "ui_windows_settings_dialog_scroll.hpp"
-#include "ui_windows_settings_dialog_paint.hpp"
-#include "ui_windows_settings_dialog_handlers.hpp"
-} // namespace settings_dlg
-
-#include "ui_windows_settings_dialog_show.hpp"
-} // namespace ui
+#include "ui_windows_settings_dialog.hpp"
 
 #include "debug.hpp"
 #include "dpi.hpp"
@@ -40,5 +23,9 @@ namespace settings_dlg {
 #include "window.hpp"
 
 namespace ui {
+using WindowHandle = HWND;
+using FontHandle = HFONT;
+using ThemeHandle = const Theme*;
+using InstanceHandle = HINSTANCE;
 #include "ui_windows_app.hpp"
 } // namespace ui
